+++ GMU-Projekt/src/matrix_operations.cpp
@@ -1,6 +1,6 @@
 #include "matrix_operations.h"
 
-void determinant(double *A, long double *d, int *exp, int matrixSize)
+void determinant(float *A, long double *d, int *exp, int matrixSize)
 {
 	*d = 1.0;
 	*exp = 0;
@@ -23,7 +23,7 @@ void determinant(double *A, long double *d, int *exp, int matrixSize)
 			// swap lines -- FOR can be parallelized
 			for (int sx = pivot; sx < matrixSize; sx++)
 			{
-				long double tmp = A[pivot*matrixSize + sx];
+				long float tmp = A[pivot*matrixSize + sx];
 				A[pivot*matrixSize + sx] = A[sy*matrixSize + sx];
 				A[sy*matrixSize + sx] = tmp;
 			}
@@ -32,7 +32,7 @@ void determinant(double *A, long double *d, int *exp, int matrixSize)
 		// FOR can be parallelized
 		for (int y = pivot + 1; y < matrixSize; y++)
 		{
-			long double multiplier = A[y*matrixSize + pivot] / A[pivot*matrixSize + pivot];
+			long float multiplier = A[y*matrixSize + pivot] / A[pivot*matrixSize + pivot];
 
 			// multiply and substract "pivot" row from "y" row -- FOR can be parallelized
 			for (int x = pivot + 1; x < matrixSize; x++)
@@ -59,7 +59,7 @@ void determinant(double *A, long double *d, int *exp, int matrixSize)
 	}
 }
 
-void gem(double *A, double *result, int matrixSize)
+void gem(float *A, float *result, int matrixSize)
 {
 	for (int i = 0; i < matrixSize; i++)
 	{
@@ -84,14 +84,14 @@ void gem(double *A, double *result, int matrixSize)
 			// swap lines -- FOR can be parallelized
 			for (int sx = pivot; sx < matrixSize; sx++)
 			{
-				long double tmp = A[pivot*matrixSize + sx];
+				long float tmp = A[pivot*matrixSize + sx];
 				A[pivot*matrixSize + sx] = -A[sy*matrixSize + sx];
 				A[sy*matrixSize + sx] = tmp;
 			}
 			result[pivot] *= -1;
 		}
 
-		double divider = A[pivot*matrixSize + pivot];
+		float divider = A[pivot*matrixSize + pivot];
 		for (int x = pivot; x < matrixSize; x++)
 		{
 			A[pivot*matrixSize + x] /= divider;
@@ -103,7 +103,7 @@ void gem(double *A, double *result, int matrixSize)
 		{
 			if (y != pivot)
 			{
-				long double multiplier = A[y*matrixSize + pivot] / A[pivot*matrixSize + pivot];
+				long float multiplier = A[y*matrixSize + pivot] / A[pivot*matrixSize + pivot];
 
 				// multiply and substract "pivot" row from "y" row -- FOR can be parallelized
 				for (int x = pivot + 1; x < matrixSize; x++)
@@ -119,7 +119,7 @@ void gem(double *A, double *result, int matrixSize)
 	}
 }
 
-void inverse(double *A, double *result, int matrixSize)
+void inverse(float *A, float *result, int matrixSize)
 {
 	// fill ones to diagonal of result matrix
 	for (int i = 0; i < matrixSize*matrixSize; i++)
@@ -144,7 +144,7 @@ void inverse(double *A, double *result, int matrixSize)
 			// swap lines -- FOR can be parallelized
 			for (int sx = pivot; sx < matrixSize; sx++)
 			{
-				long double tmp = A[pivot*matrixSize + sx];
+				long float tmp = A[pivot*matrixSize + sx];
 				A[pivot*matrixSize + sx] = -A[sy*matrixSize + sx];
 				A[sy*matrixSize + sx] = tmp;
 
@@ -152,7 +152,7 @@ void inverse(double *A, double *result, int matrixSize)
 			}
 		}
 
-		double divider = A[pivot*matrixSize + pivot];
+		float divider = A[pivot*matrixSize + pivot];
 		for (int x = pivot; x < matrixSize; x++)
 		{
 			A[pivot*matrixSize + x] /= divider;
@@ -167,7 +167,7 @@ void inverse(double *A, double *result, int matrixSize)
 		{
 			if (y != pivot)
 			{
-				long double multiplier = A[y*matrixSize + pivot] / A[pivot*matrixSize + pivot];
+				long float multiplier = A[y*matrixSize + pivot] / A[pivot*matrixSize + pivot];
 
 				// multiply and substract "pivot" row from "y" row -- FOR can be parallelized
 				for (int x = pivot + 1; x < matrixSize; x++)
